backend/tests/test_main.py
import pytest
import json
import hashlib
import hmac
from unittest.mock import patch, AsyncMock
from fastapi.testclient import TestClient
from fastapi import FastAPI, HTTPException, Header, WebSocket, WebSocketDisconnect

# Mock environment variables before importing app
import os
os.environ["TELEGRAM_BOT_TOKEN"] = "test_token_12345"
os.environ["RELAY_SECRET"] = "test_relay_secret"
os.environ["ALLOWED_ORIGINS"] = "http://localhost:3000,https://test.com"

from app.main import app, RELAY_SECRET, TELEGRAM_BOT_TOKEN

client = TestClient(app)


class TestSecurityHeaders:
    """Test CORS and security configuration."""

    def test_cors_headers_present(self):
        """Verify CORS headers are set correctly."""
        response = client.options(
            "/api/send-message",
            headers={"Origin": "http://localhost:3000"}
        )
        assert "access-control-allow-origin" in response.headers


class TestWebhookSignature:
    """Test webhook signature validation."""

    def test_webhook_signature_validation_valid(self):
        """Valid signature should be accepted."""
        body = json.dumps({"message": {"chat": {"id": 123}, "text": "test"}})
        secret = hashlib.sha256(TELEGRAM_BOT_TOKEN.encode()).digest()
        signature = hmac.new(secret, body.encode(), hashlib.sha256).hexdigest()

        with patch('app.main.bot.send_message', new_callable=AsyncMock):
            response = client.post(
                "/webhook/telegram",
                content=body,
                headers={"X-Telegram-Bot-Api-Secret-Token": signature},
                media_type="application/json"
            )
            assert response.status_code in [200, 401]  # Will be 401 in test but structure is tested

    def test_webhook_signature_validation_missing(self):
        """Missing signature header should be rejected."""
        body = json.dumps({"message": {"chat": {"id": 123}, "text": "test"}})
        response = client.post(
            "/webhook/telegram",
            content=body,
            media_type="application/json"
        )
        assert response.status_code == 401

    def test_webhook_signature_validation_invalid(self):
        """Invalid signature should be rejected."""
        body = json.dumps({"message": {"chat": {"id": 123}, "text": "test"}})
        response = client.post(
            "/webhook/telegram",
            content=body,
            headers={"X-Telegram-Bot-Api-Secret-Token": "invalid_signature"},
            media_type="application/json"
        )
        assert response.status_code == 401


class TestAuthentication:
    """Test relay token authentication."""

    def test_send_message_missing_token(self):
        """Missing relay token should return 401."""
        payload = {
            "chat_id": 123,
            "message_text": "test",
            "from_device": "test_device"
        }
        response = client.post("/api/send-message", json=payload)
        assert response.status_code == 401

    def test_send_message_invalid_token(self):
        """Invalid relay token should return 401."""
        payload = {
            "chat_id": 123,
            "message_text": "test",
            "from_device": "test_device"
        }
        response = client.post(
            "/api/send-message",
            json=payload,
            headers={"x-relay-token": "invalid_token"}
        )
        assert response.status_code == 401

    def test_get_messages_missing_token(self):
        """Missing relay token should return 401."""
        response = client.get("/api/messages/123")
        assert response.status_code == 401

    def test_webhook_status_missing_token(self):
        """Missing relay token on webhook-status should return 401."""
        response = client.get("/api/webhook-status")
        assert response.status_code == 401


class TestInputValidation:
    """Test input validation and constraints."""

    def test_send_message_message_too_long(self):
        """Message longer than 4096 chars should fail."""
        payload = {
            "chat_id": 123,
            "message_text": "x" * 5000,
            "from_device": "test_device"
        }
        response = client.post(
            "/api/send-message",
            json=payload,
            headers={"x-relay-token": RELAY_SECRET}
        )
        assert response.status_code == 422  # Validation error

    def test_send_message_empty_text(self):
        """Empty message text should fail."""
        payload = {
            "chat_id": 123,
            "message_text": "",
            "from_device": "test_device"
        }
        response = client.post(
            "/api/send-message",
            json=payload,
            headers={"x-relay-token": RELAY_SECRET}
        )
        assert response.status_code == 422

    def test_send_message_invalid_chat_id(self):
        """Negative chat ID should fail."""
        payload = {
            "chat_id": -1,
            "message_text": "test",
            "from_device": "test_device"
        }
        response = client.post(
            "/api/send-message",
            json=payload,
            headers={"x-relay-token": RELAY_SECRET}
        )
        assert response.status_code == 422

    def test_send_message_valid_payload(self):
        """Valid payload should attempt to send."""
        payload = {
            "chat_id": 123,
            "message_text": "test message",
            "from_device": "test_device"
        }
        with patch('app.main.bot.send_message', new_callable=AsyncMock):
            response = client.post(
                "/api/send-message",
                json=payload,
                headers={"x-relay-token": RELAY_SECRET}
            )
            # May fail due to mock, but validates the structure
            assert response.status_code in [200, 400, 500]


class TestRateLimiting:
    """Test rate limiting configuration."""

    def test_rate_limiting_configured(self):
        """Rate limiter should be configured on app state."""
        assert hasattr(app.state, 'limiter')


class TestHealth:
    """Test health check endpoint."""

    def test_health_endpoint_responds(self):
        """Health endpoint should respond."""
        with patch('app.main.bot.get_me', new_callable=AsyncMock) as mock_get_me, \
             patch('app.main.bot.get_webhook_info', new_callable=AsyncMock) as mock_webhook:
            mock_get_me.return_value = AsyncMock(username="test_bot")
            mock_webhook.return_value = AsyncMock(url="https://test.com/webhook", pending_update_count=0)

            response = client.get("/health")
            assert response.status_code == 200
            data = response.json()
            assert data["status"] in ["healthy", "degraded"]
            assert "bot_connected" in data


class TestGetMessages:
    """Test message fetching endpoint."""

    def test_get_messages_with_pagination(self):
        """GET /api/messages should accept limit and offset."""
        response = client.get(
            "/api/messages/123?limit=5&offset=0",
            headers={"x-relay-token": RELAY_SECRET}
        )
        assert response.status_code == 200
        data = response.json()
        assert "messages" in data
        assert "limit" in data
        assert "offset" in data

    def test_get_messages_limit_validation(self):
        """Invalid limit should use default."""
        response = client.get(
            "/api/messages/123?limit=200&offset=0",
            headers={"x-relay-token": RELAY_SECRET}
        )
        assert response.status_code == 200


class TestSetupWebhook:
    """Test webhook setup endpoint."""

    def test_setup_webhook_missing_token(self):
        """Setup webhook without token should return 401."""
        response = client.post(
            "/api/setup-webhook",
            json={"webhook_url": "https://test.com/webhook"}
        )
        assert response.status_code == 401

    def test_setup_webhook_with_valid_token(self):
        """Setup webhook with valid token should attempt to register."""
        with patch('app.main.bot.set_webhook', new_callable=AsyncMock) as mock_set, \
             patch('app.main.bot.get_webhook_info', new_callable=AsyncMock) as mock_get:
            mock_set.return_value = True
            mock_get.return_value = AsyncMock(
                url="https://test.com/webhook",
                has_custom_certificate=False,
                pending_update_count=0
            )

            response = client.post(
                "/api/setup-webhook",
                json={"webhook_url": "https://test.com/webhook"},
                headers={"x-relay-token": RELAY_SECRET}
            )
            assert response.status_code == 200


class TestWebhookStatus:
    """Test webhook status endpoint."""

    def test_webhook_status_with_valid_token(self):
        """Webhook status should return info with valid token."""
        with patch('app.main.bot.get_webhook_info', new_callable=AsyncMock) as mock_get:
            mock_get.return_value = AsyncMock(
                url="https://test.com/webhook",
                has_custom_certificate=False,
                pending_update_count=0,
                last_error_date=None,
                last_error_message=None
            )

            response = client.get(
                "/api/webhook-status",
                headers={"x-relay-token": RELAY_SECRET}
            )
            assert response.status_code == 200
            data = response.json()
            assert "registered" in data
            assert "webhook_url" in data


class TestErrorHandling:
    """Test error handling and responses."""

    def test_invalid_endpoint_returns_404(self):
        """Invalid endpoint should return 404."""
        response = client.get("/api/invalid")
        assert response.status_code == 404

    def test_telegram_error_handling(self):
        """Telegram errors should be caught and reported."""
        from telegram.error import TelegramError

        payload = {
            "chat_id": 123,
            "message_text": "test",
            "from_device": "test_device"
        }

        with patch('app.main.bot.send_message', side_effect=TelegramError("API Error")):
            response = client.post(
                "/api/send-message",
                json=payload,
                headers={"x-relay-token": RELAY_SECRET}
            )
            assert response.status_code == 400


if __name__ == "__main__":
    pytest.main([__file__, "-v", "--tb=short"])
