backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot>=21.0,<22.0
requests>=2.31.0,<3.0.0
slowapi==0.1.8
python-dotenv>=1.0.0,<2.0.0

# Testing & Security
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.25.0,<1.0.0
bandit>=1.7.5,<2.0.0
black>=23.10.0,<24.0.0
ruff>=0.1.0,<1.0.0
